messages are successfully rerendering page
