+++ server/requirements.txt
@@ -1,6 +1,8 @@
 click==8.0.3
 Flask==2.0.2
+Flask-Cors==3.0.10
 itsdangerous==2.0.1
 Jinja2==3.0.3
 MarkupSafe==2.0.1
+six==1.16.0
 Werkzeug==2.0.2
